+++ lua/config/bufferline.lua
@@ -6,7 +6,7 @@ bufferline.setup({
     options = {
         offsets = {
             {
-                filetype = "NvimTree",
+                filetype = "neo-tree",
                 text = "File Explorer",
                 highlight = "Directory",
                 text_align = "left"