+++ lua/config/nvim-tree.lua
@@ -1,7 +1,0 @@
--- lua/config/nvim-tree.lua
-
-local browser = require('nvim-tree')
-
-browser.setup({
-    disable_netrw = true,
-})