+++ lua/plugins/nvim-tree.lua
@@ -1,8 +1,0 @@
--- lua/plugins/nvim-tree.lua
-
-local use = require('packer').use
-
-use({
-    'nvim-tree/nvim-tree.lua',
-    requires = { 'nvim-tree/nvim-web-devicons', opt = true },
-})